+++ src/index.ts
@@ -44,6 +44,7 @@ import { GenLiteHealthRegenerationPlugin } from './plugins/genlite-health-regene
 import { GenLiteFPSCounter } from "./plugins/genlite-fps.plugin";
 import { GenLiteEnhancedContextMenu } from "./plugins/genlite-enhanced-context-menu.plugin";
 import { GenLiteQuestPlugin } from "./plugins/genlite-quest.plugin";
+import { GenLiteEnhancedBanking } from "./plugins/genlite-enhanced-banking.plugin";
 
 declare const GM_getResourceText: (s: string) => string;
 
@@ -227,6 +228,7 @@ let isInitialized = false;
         await genlite.pluginLoader.addPlugin(GenLiteFPSCounter);
         await genlite.pluginLoader.addPlugin(GenLiteEnhancedContextMenu);
         await genlite.pluginLoader.addPlugin(GenLiteQuestPlugin);
+        await genlite.pluginLoader.addPlugin(GenLiteEnhancedBanking);
 
         /** post init things */
         // await document['GenLiteDatabasePlugin'].postInit();
